backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
chromadb>=1.0.0
FlagEmbedding>=1.3.0
anthropic>=0.40.0
httpx>=0.27.0
extract-msg>=0.50.0
